+++ R/utils-misc.R
@@ -1,22 +1,3 @@
-has_gs <- function() {
-    has_cmd("ghostscript", tools::find_gs_cmd)
-}
-
-gs <- function() {
-    get_cmd("ghostscript", tools::find_gs_cmd)
-}
-
-has_cmd <- function(name, cmd_fn = function() Sys.which(name)) {
-    cmd_fn() != ""
-}
-
-get_cmd <- function(name, cmd_fn = function() Sys.which(name)) {
-    cmd <- cmd_fn()
-    if (cmd == "")
-        abort(sprintf("Can't find system dependency `%s` on PATH", name))
-    cmd
-}
-
 assert_suggested <- function(package) {
     calling_fn <- deparse(sys.calls()[[sys.nframe()-1]])
     if (!requireNamespace(package, quietly = TRUE)) {